+++ test/integration/runTestsFromPath.test.ts
@@ -7,6 +7,7 @@ import * as chaiSubset from "chai-subset";
 import * as flatMap from "array.prototype.flatmap";
 import * as yn from "yn";
 import { CLIEngine } from "eslint";
+import * as dotEnv from "dotenv"
 import EntityFileToJson from "../utils/EntityFileToJson";
 import { createDriver, dataCollectionPhase } from "../../src/Engine";
 import * as GTU from "../utils/GeneralTestUtils";
@@ -15,7 +16,7 @@ import IConnectionOptions from "../../src/IConnectionOptions";
 import modelCustomizationPhase from "../../src/ModelCustomization";
 import modelGenerationPhase from "../../src/ModelGeneration";
 
-require("dotenv").config();
+dotEnv.config();
 
 flatMap.shim();
 chai.use(chaiSubset);
@@ -37,22 +38,57 @@ describe("TypeOrm examples", async () => {
     const testPartialPath = "test/integration/examples";
     await runTestsFromPath(testPartialPath, false);
 });
+describe("Filtering tables", async () => {
+    const testPartialPath = "test/integration/examples";
+    it("skipTables",async ()=>{
+        const dbDrivers: string[] = GTU.getEnabledDbDrivers();
+        const modelGenerationPromises = dbDrivers.map(async dbDriver => {
+            const {
+                generationOptions,
+                driver,
+                connectionOptions
+            } = await prepareTestRuns(testPartialPath, "sample2-one-to-one", dbDriver);
+            const dbModel = await dataCollectionPhase(
+                        driver,
+                        {...connectionOptions,
+                        skipTables:["Post"]},
+                        generationOptions
+                    );
+            expect(dbModel.length).to.equal(6);
+            // eslint-disable-next-line no-unused-expressions
+            expect(dbModel.find(x=>x.sqlName==="Post")).to.be.undefined;
+        });
+        await Promise.all(modelGenerationPromises);
+    });
+    it("onlyTables",async ()=>{
+        const dbDrivers: string[] = GTU.getEnabledDbDrivers();
+        const modelGenerationPromises = dbDrivers.map(async dbDriver => {
+            const {
+                generationOptions,
+                driver,
+                connectionOptions
+            } = await prepareTestRuns(testPartialPath, "sample2-one-to-one", dbDriver);
+            const dbModel = await dataCollectionPhase(
+                        driver,
+                        {...connectionOptions,
+                        onlyTables:["Post"]},
+                        generationOptions
+                    );
+            expect(dbModel.length).to.equal(1);
+            // eslint-disable-next-line no-unused-expressions
+            expect(dbModel.find(x=>x.sqlName==="Post")).to.not.be.undefined;
+        });
+        await Promise.all(modelGenerationPromises);
+    })
+
+});
 
 async function runTestsFromPath(
     testPartialPath: string,
     isDbSpecific: boolean
 ) {
-    const resultsPath = path.resolve(process.cwd(), `output`);
-    if (!fs.existsSync(resultsPath)) {
-        fs.mkdirSync(resultsPath);
-    }
     const dbDrivers: string[] = GTU.getEnabledDbDrivers();
-    dbDrivers.forEach(dbDriver => {
-        const newDirPath = path.resolve(resultsPath, dbDriver);
-        if (!fs.existsSync(newDirPath)) {
-            fs.mkdirSync(newDirPath);
-        }
-    });
+    createOutputDirs(dbDrivers);
     const files = fs.readdirSync(path.resolve(process.cwd(), testPartialPath));
     if (isDbSpecific) {
         await runTest(dbDrivers, testPartialPath, files);
@@ -61,6 +97,19 @@ async function runTestsFromPath(
             runTestForMultipleDrivers(folder, dbDrivers, testPartialPath);
         });
     }
+}
+
+function createOutputDirs(dbDrivers: string[]) {
+    const resultsPath = path.resolve(process.cwd(), `output`);
+    if (!fs.existsSync(resultsPath)) {
+        fs.mkdirSync(resultsPath);
+    }
+    dbDrivers.forEach(dbDriver => {
+        const newDirPath = path.resolve(resultsPath, dbDriver);
+        if (!fs.existsSync(newDirPath)) {
+            fs.mkdirSync(newDirPath);
+        }
+    });
 }
 
 function runTestForMultipleDrivers(
@@ -349,7 +398,8 @@ async function prepareTestRuns(
                         databaseType: "mysql",
                         schemaName: "ignored",
                         ssl: yn(process.env.MYSQL_SSL, { default: false }),
-                        skipTables: []
+                        skipTables: [],
+                        onlyTables: [],
                     };
                     break;
                 case "mariadb":
@@ -362,7 +412,8 @@ async function prepareTestRuns(
                         databaseType: "mariadb",
                         schemaName: "ignored",
                         ssl: yn(process.env.MARIADB_SSL, { default: false }),
-                        skipTables: []
+                        skipTables: [],
+                        onlyTables: []
                     };
                     break;